Auth-and-Management/backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Authlib==1.3.1
Flask-Mail==0.9.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.7.0
flask-cors==4.0.1
PyMySQL==1.1.1
cryptography==42.0.5
python-dotenv==1.0.1
bcrypt==4.1.3
requests==2.31.0
